+++ speech_to_text/ios/Classes/SwiftSpeechToTextPlugin.swift
@@ -111,11 +111,16 @@ public class SwiftSpeechToTextPlugin: NSObject, FlutterPlugin {
             initialize( result )
         case SwiftSpeechToTextMethods.listen.rawValue:
             guard let argsArr = call.arguments as? Dictionary<String,AnyObject>,
-                let partialResults = argsArr["partialResults"] as? Bool, let onDevice = argsArr["onDevice"] as? Bool, let listenModeIndex = argsArr["listenMode"] as? Int, let sampleRate = argsArr["sampleRate"] as? Int
+                let partialResults = argsArr["partialResults"] as? Bool, 
+                    let onDevice = argsArr["onDevice"] as? Bool,
+                    let listenModeIndex = argsArr["listenMode"] as? Int,
+                    let sampleRate = argsArr["sampleRate"] as? Int,
+                    let autoPunctuation = argsArr["autoPunctuation"] as? Bool,
+                    let enableHaptics = argsArr["enableHaptics"] as? Bool
                 else {
                     DispatchQueue.main.async {
                         result(FlutterError( code: SpeechToTextErrors.missingOrInvalidArg.rawValue,
-                                             message:"Missing arg partialResults, onDevice, listenMode, and sampleRate are required",
+                                             message:"Missing arg partialResults, onDevice, listenMode, autoPunctuatio, enableHaptics and sampleRate are required",
                                              details: nil ))
                     }
                     return
@@ -133,7 +138,7 @@ public class SwiftSpeechToTextPlugin: NSObject, FlutterPlugin {
                 return
             }
             
-            listenForSpeech( result, localeStr: localeStr, partialResults: partialResults, onDevice: onDevice, listenMode: listenMode, sampleRate: sampleRate )
+            listenForSpeech( result, localeStr: localeStr, partialResults: partialResults, onDevice: onDevice, listenMode: listenMode, sampleRate: sampleRate, autoPunctuation: autoPunctuation, enableHaptics: enableHaptics )
         case SwiftSpeechToTextMethods.stop.rawValue:
             stopSpeech( result )
         case SwiftSpeechToTextMethods.cancel.rawValue:
@@ -348,7 +353,8 @@ public class SwiftSpeechToTextPlugin: NSObject, FlutterPlugin {
         stopping = false
     }
     
-    private func listenForSpeech( _ result: @escaping FlutterResult, localeStr: String?, partialResults: Bool, onDevice: Bool, listenMode: ListenMode, sampleRate: Int ) {
+    private func listenForSpeech( _ result: @escaping FlutterResult, localeStr: String?, partialResults: Bool, 
+                                  onDevice: Bool, listenMode: ListenMode, sampleRate: Int, autoPunctuation: Bool, enableHaptics: Bool ) {
         if ( nil != currentTask || listening ) {
             sendBoolResult( false, result );
             return
@@ -381,6 +387,9 @@ public class SwiftSpeechToTextPlugin: NSObject, FlutterPlugin {
             }
             try self.audioSession.setMode(AVAudioSession.Mode.default)
             try self.audioSession.setActive(true, options: .notifyOthersOnDeactivation)
+            if #available(iOS 13.0, *) {
+                try self.audioSession.setAllowHapticsAndSystemSoundsDuringRecording(enableHaptics)
+            }
             if let sound = listeningSound {
                 self.onPlayEnd = {()->Void in
                     if ( !self.failedListen ) {
@@ -417,7 +426,9 @@ public class SwiftSpeechToTextPlugin: NSObject, FlutterPlugin {
             default:
                 break
             }
-            
+            if #available(iOS 16.0, *) {
+                currentRequest.addsPunctuation = autoPunctuation
+            }
             self.currentTask = self.recognizer?.recognitionTask(with: currentRequest, delegate: self )
             let recordingFormat = inputNode?.outputFormat(forBus: self.busForNodeTap)
             let theSampleRate = audioSession.sampleRate